+++ plone/subrequest/__init__.py
@@ -1,5 +1,4 @@
-import re
-
+# -*- coding: utf-8 -*-
 from AccessControl import getSecurityManager
 from AccessControl.SecurityManagement import setSecurityManager
 from Acquisition import aq_base
@@ -9,18 +8,20 @@ from ZPublisher.Publish import missing_name
 from ZPublisher.mapply import mapply
 from cStringIO import StringIO
 from logging import getLogger
+from plone.subrequest.interfaces import ISubRequest
+from plone.subrequest.subresponse import SubResponse
 from posixpath import normpath
+from urllib import unquote  # Python2.4 does not have urlparse.unquote
 from urlparse import urlsplit, urljoin
-from urllib import unquote # Python2.4 does not have urlparse.unquote
-from zope.interface import alsoProvides
 from zope.globalrequest import getRequest, setRequest
+from zope.interface import alsoProvides
+import re
+
 try:
     from zope.site.hooks import getSite, setSite
 except ImportError:
     from zope.app.component.hooks import getSite, setSite
 
-from plone.subrequest.subresponse import SubResponse
-from plone.subrequest.interfaces import ISubRequest
 
 __all__ = ['subrequest', 'SubResponse']
 
@@ -31,7 +32,7 @@ CONDITIONAL_HEADERS = [
     'HTTP_IF_MATCH',
     'HTTP_IF_NONE_MATCH',
     'HTTP_IF_RANGE',
-    'HTTP_RANGE', # Not strictly a conditional header, but scrub it anyway
+    'HTTP_RANGE',  # Not strictly a conditional header, but scrub it anyway
     ]
 
 OTHER_IGNORE = set([
@@ -55,13 +56,15 @@ OTHER_IGNORE_RE = re.compile(r'^(?:BASE|URL)\d+$')
 
 logger = getLogger("plone.subrequest")
 
+
 def subrequest(url, root=None, stdout=None):
     assert url is not None, "You must pass a url"
     if isinstance(url, unicode):
         url = url.encode('utf-8')
     _, _, path, query, _ = urlsplit(url)
     parent_request = getRequest()
-    assert parent_request is not None, "Unable to get request, perhaps zope.globalrequest is not configured."
+    assert parent_request is not None, \
+        "Unable to get request, perhaps zope.globalrequest is not configured."
     parent_site = getSite()
     security_manager = getSecurityManager()
     parent_app = parent_request.PARENTS[-1]
@@ -71,7 +74,9 @@ def subrequest(url, root=None, stdout=None):
         if vurl_parts is not None:
             # Use the virtual host root
             path_past_root = unquote(vurl_parts[-1])
-            root_path = normpath(parent_request['PATH_INFO']).rstrip('/')[:-len(path_past_root) or None]
+            root_path = normpath(
+                parent_request['PATH_INFO']
+            ).rstrip('/')[:-len(path_past_root) or None]
             if root is None:
                 path = root_path + path
             else:
@@ -92,9 +97,13 @@ def subrequest(url, root=None, stdout=None):
         path = normpath(path)
     request = parent_request.clone()
     for name, parent_value in parent_request.other.items():
-        if name in OTHER_IGNORE or OTHER_IGNORE_RE.match(name) or name.startswith('_'):
+        if name in OTHER_IGNORE \
+           or OTHER_IGNORE_RE.match(name) \
+           or name.startswith('_'):
             continue
         request.other[name] = parent_value
+    for key, value in parent_request.response.cookies.items():
+        request.cookies[key] = value['value']
     request['PARENT_REQUEST'] = parent_request
     alsoProvides(request, ISubRequest)
     try:
@@ -104,9 +113,9 @@ def subrequest(url, root=None, stdout=None):
         request['PARENTS'] = [app]
         response = request.response
         response.__class__ = SubResponse
-        response.stderr = None # only used on retry it seems
+        response.stderr = None  # only used on retry it seems
         if stdout is None:
-            stdout = StringIO() # It might be possible to optimize this
+            stdout = StringIO()  # It might be possible to optimize this
         response.stdout = stdout
         environ = request.environ
         environ['PATH_INFO'] = path
@@ -138,4 +147,3 @@ def subrequest(url, root=None, stdout=None):
         setRequest(parent_request)
         setSite(parent_site)
         setSecurityManager(security_manager)
-